+++ app/javascript/pageflip.js
@@ -1,0 +1,14 @@
+import { PageFlip } from "page-flip";
+
+document.addEventListener("DOMContentLoaded", () => {
+  const elm = document.getElementById("book");
+
+  if (elm) {
+    const pageFlip = new PageFlip(document.getElementById('book'), {
+      width: 400,
+      height: 600,
+    });
+
+    pageFlip.loadFromHTML(document.querySelectorAll('.my-page'));
+  }
+});
